# admin/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Gym Genius application",
        "title": "Gym Genius API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/exercises/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exercises"
                ],
                "summary": "Get an exercise by id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ExerciseRead"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exercises"
                ],
                "summary": "Delete an exercise by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Exercise ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exercise deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/models.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Exercise Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exercises"
                ],
                "summary": "Update an existing exercise by ID",
                "parameters": [
                    {
                        "description": "Exercise update payload",
                        "name": "exercise",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ExerciseUpdate"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Exercise ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Exercise Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/statistics/active-users": {
            "get": {
                "description": "Returns the count of distinct active users grouped by specified time intervals between start and end dates.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Get number of active users in time intervals",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date/time in RFC3339 format (e.g., 2025-06-01T00:00:00Z)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date/time in RFC3339 format (e.g., 2025-06-10T00:00:00Z)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"24h\"",
                        "description": "Time interval step duration, format: \u003cnumber\u003e\u003cunit\u003e where unit is 'h' for hours or 'd' for days (e.g., '24h', '7d')",
                        "name": "step",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of interval results containing start time, end time, and user count",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.intervalResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input parameters (missing, malformed, or invalid date range)",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error while processing the request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/{id}/activity": {
            "get": {
                "description": "Retrieve monthly activity statistics for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Get user activity statistics",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Monthly activity statistics",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/handlers.MonthlyStat"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.MonthlyStat": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "month": {
                    "type": "string"
                }
            }
        },
        "handlers.intervalResult": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "A descriptive error message"
                }
            }
        },
...
        "models.ExerciseRead": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Push-ups are a basic exercise that works the chest, shoulders, and triceps."
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "muscle_groups": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "chest",
                        "back",
                        "triceps"
                    ]
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "description": "URL to the exercise image",
                    "type": "string",
                    "example": "https://example.com/image.jpg"
                }
            }
        },
...
        "models.ExerciseUpdate": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Push-ups are a basic exercise that works the chest, shoulders, and triceps."
                },
                "image_path": {
                    "description": "URL to the exercise image",
                    "type": "string",
                    "example": "https://example.com/image.jpg"
                },
                "muscle_group": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "chest",
                        "back",
                        "triceps"
                    ]
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Descriptive message"
                }
            }
        },
...
    }
}